WIP: Use CMake package in data-sealing sample
Most of the complexity in this CMake build is because the project is
structured in a way that is not conducive to CMake, and I did not want
to change the GNU Make build system. One issue is that the generated EDL
files are expected in the source folder `common`, which poses a problem
because the generated output of `add_custom_command` can only be used in
CMake commands in the same folder (or subfolder). So `common` being next
to `enc1` etc. makes this awkward. A similar story plays out for the
generated keys and signed enclaves.

If we deleted the GNU Make build and were willing to move a few things
around, this could be a lot cleaner. As it is, it works:

```
...
Host: operation succeeded: completed as expected
Host: Terminating enclaves
Host: enclave terminated.
Host: enclave terminated.
Host: enclave terminated.
Host: Sample completed successfully.
```