Exclude OE_API_VERSION=2 from CMake package
The samples do not expect this to be defined, and we did not intend to
define it for all users of the SDK (yet). However, it should be
internally public because our own build system expects it to be defined.
So we constrain it to our build system but not our package using a
generator expression to exclude it from the install interface.